requirements.txt
Flask==2.2.5
flasgger==0.9.5
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Werkzeug==3.0.6
PyMySQL==1.1.0
python-dotenv
flasgger==0.9.5
gunicorn
cryptography
flask_cors